+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 0.1.0
 - Use setup instead of require to collect bundled gems
 - Added `withoutOpal` option
+- Local compile-opal script